fix(view-map): Now use `rl-floorplan` component to render maps.
An SVG-as-JSON object is now used to generate the SVG floorplan meaning the MapElement `path` is no
longer needed, it is all in the JSON data.  This means that there is no need for MapElement and
MapElementDetail.  So MapElementDetail is now MapElement.

BREAKING CHANGE: Breaking changes galore.  Interfaces for MapElement changed.  No longer uses
Elements and Details, just Details (and they're now called Elements for clarity).